Split up spec support code into separate files
