+++ spec/support/database_cleaner.rb
@@ -1,0 +1,13 @@
+require 'database_cleaner'
+
+DatabaseCleaner.strategy = :transaction
+
+class Minitest::Spec
+  before :each do
+    DatabaseCleaner.start
+  end
+
+  after :each do
+    DatabaseCleaner.clean
+  end
+end
