+++ spec/support/rack_test.rb
@@ -1,0 +1,9 @@
+require 'rack/test'
+
+class Minitest::Spec
+  include Rack::Test::Methods
+
+  def app
+    Sinatra::Application
+  end
+end
